12 03 2024
Having to spend a lot of time cleaning the STATION.CSV

I thought I would have been able to use the import export wizard like before but the address field in the STATION table has to be clarified with commas.

Previously on other problems using the space delimiter on the import export wizard was enough

Thankfully, there are only 499 rows.

This will probably take me a few pomodori.  A pomodoro is an uninterrupted 25 minute block of focused study time.