12 15 2024
I have gotten closer to an answer but I still need lots of work to do.

Not sure where to go next except to re-read the problem again.

There much be some clue as to the order of the nested functions.

I am happy that I am getting more practice with these nested functions.